requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-multipart==0.0.9
google-generativeai==0.8.3
pgvector==0.3.2
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
pydantic-settings==2.5.2
httpx==0.27.2
PyJWT==2.9.0
